Refactor and redirect translation for ToolKeyItems
