+++ app/components/ToolKeyItems/KeyItemListItem.js
@@ -1,0 +1,4 @@
+import styled from 'styled-components';
+export default styled.div`
+  margin: 10px 0;
+`;
