+++ app/containers/ToolKeyItems/staticText.js
@@ -1,0 +1,34 @@
+export default {
+  tactic: {
+    id: 'module.key-tactic',
+    defaultMessage: 'Key Tactic',
+  },
+  tactics: {
+    id: 'module.key-tactics',
+    defaultMessage: 'Key Tactics',
+  },
+  theory: {
+    id: 'module.key-theory',
+    defaultMessage: 'Key Theory',
+  },
+  theories: {
+    id: 'module.key-theories',
+    defaultMessage: 'Key Theories',
+  },
+  methodology: {
+    id: 'module.key-methodology',
+    defaultMessage: 'Key Methodology',
+  },
+  methodologies: {
+    id: 'module.key-methodologies',
+    defaultMessage: 'Key Methodologies',
+  },
+  principle: {
+    id: 'module.key-principle',
+    defaultMessage: 'Key Principle',
+  },
+  principles: {
+    id: 'module.key-principles',
+    defaultMessage: 'Key Principles',
+  }
+}
