+++ app/containers/ToolPage/ToolPageMain.js
@@ -25,7 +25,7 @@ import { ToolMainArea,
 import ToolHowToUse from 'containers/ToolHowToUse';
 import ToolWhyItWorked from 'components/ToolWhyItWorked';
 import ToolWhyItFailed from 'components/ToolWhyItFailed';
-import ToolKeyItems from 'components/ToolKeyItems';
+import ToolKeyItems from 'containers/ToolKeyItems';
 import ToolEpigraph from 'components/ToolEpigraph';
 import ToolWithPullQuote from 'components/ToolWithPullQuote';
 // import { makeSelectToolById } from 'containers/Tool/selectors';
